+++ contracts/zero-ex/test/lib_signature_test.ts
@@ -2,8 +2,7 @@ import { blockchainTests, expect } from '@0x/contracts-test-utils';
 import { hexUtils } from '@0x/utils';
 import * as ethjs from 'ethereumjs-util';
 
-import { SignatureValidationError, SignatureValidationErrorCodes } from '../src/revert_errors';
-import { eip712SignHashWithKey, ethSignHashWithKey, SignatureType } from '../src/signature_utils';
+import { eip712SignHashWithKey, ethSignHashWithKey, RevertErrors, SignatureType } from '@0x/protocol-utils';
 
 import { artifacts } from './artifacts';
 import { TestLibSignatureContract } from './wrappers';
@@ -58,7 +57,10 @@ blockchainTests.resets('LibSignature library', env => {
                 v: 1,
             };
             return expect(testLib.getSignerOfHash(hash, sig).callAsync()).to.revertWith(
-                new SignatureValidationError(SignatureValidationErrorCodes.BadSignatureData, hash),
+                new RevertErrors.Signatures.SignatureValidationError(
+                    RevertErrors.Signatures.SignatureValidationErrorCodes.BadSignatureData,
+                    hash,
+                ),
             );
         });
 
@@ -69,7 +71,10 @@ blockchainTests.resets('LibSignature library', env => {
                 r: '0xfffffffffffffffffffffffffffffffebaaedce6af48a03bbfd25e8cd0364141',
             };
             return expect(testLib.getSignerOfHash(hash, sig).callAsync()).to.revertWith(
-                new SignatureValidationError(SignatureValidationErrorCodes.BadSignatureData, hash),
+                new RevertErrors.Signatures.SignatureValidationError(
+                    RevertErrors.Signatures.SignatureValidationErrorCodes.BadSignatureData,
+                    hash,
+                ),
             );
         });
 
@@ -80,7 +85,10 @@ blockchainTests.resets('LibSignature library', env => {
                 signatureType: SignatureType.Illegal,
             };
             return expect(testLib.getSignerOfHash(hash, sig).callAsync()).to.revertWith(
-                new SignatureValidationError(SignatureValidationErrorCodes.Illegal, hash),
+                new RevertErrors.Signatures.SignatureValidationError(
+                    RevertErrors.Signatures.SignatureValidationErrorCodes.Illegal,
+                    hash,
+                ),
             );
         });
 
@@ -91,7 +99,10 @@ blockchainTests.resets('LibSignature library', env => {
                 signatureType: SignatureType.Invalid,
             };
             return expect(testLib.getSignerOfHash(hash, sig).callAsync()).to.revertWith(
-                new SignatureValidationError(SignatureValidationErrorCodes.AlwaysInvalid, hash),
+                new RevertErrors.Signatures.SignatureValidationError(
+                    RevertErrors.Signatures.SignatureValidationErrorCodes.AlwaysInvalid,
+                    hash,
+                ),
             );
         });
     });