+++ packages/protocol-utils/src/constants.ts
@@ -1,0 +1,1 @@
+export const ETH_TOKEN_ADDRESS = '0xeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee';