+++ contracts/erc20/CHANGELOG.md
@@ -5,6 +5,14 @@ Edit the package's CHANGELOG.json file only.
 
 CHANGELOG
 
+## v3.2.12 - _December 9, 2020_
+
+    * Dependencies updated
+
+## v3.2.11 - _December 7, 2020_
+
+    * Dependencies updated
+
 ## v3.2.10 - _December 3, 2020_
 
     * Dependencies updated