+++ contracts/multisig/CHANGELOG.md
@@ -5,6 +5,14 @@ Edit the package's CHANGELOG.json file only.
 
 CHANGELOG
 
+## v4.1.18 - _December 9, 2020_
+
+    * Dependencies updated
+
+## v4.1.17 - _December 7, 2020_
+
+    * Dependencies updated
+
 ## v4.1.16 - _December 3, 2020_
 
     * Dependencies updated