+++ packages/asset-swapper/CHANGELOG.md
@@ -5,6 +5,22 @@ Edit the package's CHANGELOG.json file only.
 
 CHANGELOG
 
+## v5.4.2 - _December 9, 2020_
+
+    * Dependencies updated
+
+## v5.4.1 - _December 7, 2020_
+
+    * Dependencies updated
+
+## v5.4.0 - _December 7, 2020_
+
+    * Add `takerAssetToEthRate` and `makerAssetToEthRate` to swap quote response (#49)
+
+## v5.3.1 - _December 3, 2020_
+
+    * Dependencies updated
+
 ## v5.3.0 - _December 3, 2020_
 
     * Added Crypto.com (#43)